requirements.txt
flask 
pdfplumber 
python-docx 
transformers 
sentence-transformers 
torch